Move handlers_*.dart to handlers/*.dart
