+++ app/lib/frontend/handlers.dart
@@ -19,13 +19,13 @@ import '../shared/urls.dart' as urls;
 import '../shared/utils.dart';
 
 import 'backend.dart';
-import 'handlers_admin.dart';
-import 'handlers_atom_feed.dart';
-import 'handlers_custom_api.dart';
-import 'handlers_documentation.dart';
-import 'handlers_landing.dart';
-import 'handlers_misc.dart';
-import 'handlers_redirects.dart';
+import 'handlers/admin.dart';
+import 'handlers/atom_feed.dart';
+import 'handlers/custom_api.dart';
+import 'handlers/documentation.dart';
+import 'handlers/landing.dart';
+import 'handlers/misc.dart';
+import 'handlers/redirects.dart';
 import 'models.dart';
 import 'search_service.dart';
 import 'static_files.dart';
