+++ app/lib/frontend/handlers/admin.dart
@@ -8,11 +8,11 @@ import 'dart:async';
 
 import 'package:shelf/shelf.dart' as shelf;
 
-import '../shared/handlers.dart';
+import '../../shared/handlers.dart';
 
-import 'backend.dart';
-import 'models.dart';
-import 'templates.dart';
+import '../backend.dart';
+import '../models.dart';
+import '../templates.dart';
 
 /// Handles requests for /authorized
 shelf.Response authorizedHandler(_) =>
